requirements.txt
fastapi==0.135.3
Jinja2==3.1.6
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.33.1
spotipy==2.26.0
uvicorn==0.42.0